refactor parsers and propagators, add w3c and aws support
